Update: add feature kv_unstable.
