+++ lib.rs
@@ -50,8 +50,16 @@
 
 extern crate log;
 
+#[cfg(feature = "kv_unstable")]
+mod kv;
+
 use std::{fmt, io};
-use slog::{Level, KV, b, Record, Serializer};
+
+#[cfg(not(feature = "kv_unstable"))]
+use slog::{b, Level, KV};
+
+#[cfg(feature = "kv_unstable")]
+use slog::{b, Level, Record, Serializer, KV};
 
 struct Logger;
 
@@ -84,6 +92,22 @@ impl log::Log for Logger {
         true
     }
 
+    #[cfg(not(feature = "kv_unstable"))]
+    fn log(&self, r: &log::Record) {
+        let level = log_to_slog_level(r.metadata().level());
+
+        let args = r.args();
+        let target = r.target();
+        let location = &record_as_location(r);
+        let s = slog::RecordStatic {
+            location,
+            level,
+            tag: target,
+        };
+        slog_scope::with_logger(|logger| logger.log(&slog::Record::new(&s, args, b!())))
+    }
+
+    #[cfg(feature = "kv_unstable")]
     fn log(&self, r: &log::Record) {
         let level = log_to_slog_level(r.metadata().level());
 
@@ -96,51 +120,12 @@ impl log::Log for Logger {
             level,
             tag: target,
         };
-        let mut visitor = Visitor::new();
+        let mut visitor = kv::Visitor::new();
         key_values.visit(&mut visitor).unwrap();
         slog_scope::with_logger(|logger| logger.log(&slog::Record::new(&s, args, b!(visitor))))
     }
 
     fn flush(&self) {}
-}
-
-struct Visitor {
-    kvs: Vec<(String, String)>,
-}
-
-impl Visitor {
-    pub fn new() -> Self {
-        Self { kvs: vec![] }
-    }
-}
-
-impl<'kvs, 'a> log::kv::Visitor<'kvs> for Visitor {
-    fn visit_pair(
-        &mut self,
-        key: log::kv::Key<'kvs>,
-        val: log::kv::Value<'kvs>,
-    ) -> Result<(), log::kv::Error> {
-        let key = key.to_string();
-        if let Some(val) = val.to_borrowed_str() {
-            let val = val.to_string();
-            self.kvs.push((key, val));
-        }
-        Ok(())
-    }
-}
-
-fn string_to_static_str(s: String) -> &'static str {
-    Box::leak(s.into_boxed_str())
-}
-
-impl slog::KV for Visitor {
-    fn serialize(&self, _record: &Record, serializer: &mut dyn Serializer) -> slog::Result {
-        for (key, val) in &self.kvs {
-            let key = string_to_static_str(key.to_owned());
-            serializer.emit_str(key, val.as_str())?;
-        }
-        Ok(())
-    }
 }
 
 /// Register `slog-stdlog` as `log` backend.
